+++ src/Curiosity.Tools.Web/ModelBinders/Decimal/InvariantDecimalModelBinder.cs
@@ -7,15 +7,20 @@ using Microsoft.Extensions.Logging;
 
 namespace Curiosity.Tools.Web.ModelBinders
 {
+    /// <summary>
+    /// Binder of decimal data type. Uses invariant culture to bind decimal.
+    /// </summary>
     public class InvariantDecimalModelBinder : IModelBinder
     {
         private readonly SimpleTypeModelBinder _baseBinder;
 
+        /// <inheritdoc cref="InvariantDecimalModelBinder"/>
         public InvariantDecimalModelBinder(Type modelType, ILoggerFactory loggerFactory)
         {
             _baseBinder = new SimpleTypeModelBinder(modelType, loggerFactory);
         }
 
+        /// <inheritdoc />
         public Task BindModelAsync(ModelBindingContext bindingContext)
         {
             if (bindingContext == null) throw new ArgumentNullException(nameof(bindingContext));