+++ src/Curiosity.Tools.Web/ModelBinders/TrimString/TrimStringSystemJsonConverter.cs
@@ -14,11 +14,13 @@ namespace Curiosity.Tools.Web.ModelBinders
     /// </remarks>
     public class TrimStringSystemJsonConverter : JsonConverter<string?>
     {
+        /// <inheritdoc />
         public override string? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
         {
             return reader.GetString()?.Trim();
         }
 
+        /// <inheritdoc />
         public override void Write(Utf8JsonWriter writer, string? value, JsonSerializerOptions options)
         {
             writer.WriteStringValue(value?.Trim());